Moved blacklistWord to RedactedConversationBuilder for redacting words.
